+++ src/ts/dom/frontend/FrontendUI.tsx
@@ -13,31 +13,7 @@ import { IGlobalLibs } from "../../components/interfaces";
 import { SceneContainer } from "../../components/SceneContainer";
 import { JSXInternal } from "preact/src/jsx";
 import { ChapterIntro } from "./ChapterIntro";
-import { Test } from "./Test";
-// import { TextCutsceneUI } from "./TextCutsceneUI";
-
-// const introText = `
-// Year: 2663
-
-// You are financially bankrupt.
-// After making several losses in your energy business, developing a
-// next generation of enhanced nuclear reactors, you stay left incapacitatded. Your
-// tech partners have already disappeared and two handful of clients demand
-// their investments back. Well, those 1.7 million won't gain from nowhere and
-// your private accounts have already been seized. Before the authorities of
-// the Clans Union would accuse you of fraudulent activities it's better to disappear
-// for a while. Using jump ships
-// would cause a lot of beaurocratic waves so you make your way and bribe
-// some freighter captain with your last credits.
-
-// After spending weeks in a dirty cabin and ear pinching humming
-// noise of the main engine you try to make your way.
-
-// Before entering the Skagerrak the captain dedicates an old cockle boat to you.
-// This is all you get to make the rest of your way on your own. At least there is
-// enough oxygen in your tanks to reach your destination:
-// Lagertha Electrolyte Research Base in the cold waters of the Barents Sea. It should
-// only be a journey of a few more days ...`;
+import { Params } from "../../utils/Params";
 
 export interface IFrontendUIBaseProps {
   sceneContainer: SceneContainer;
@@ -45,13 +21,13 @@ export interface IFrontendUIBaseProps {
 }
 
 interface IAppProps extends IFrontendUIBaseProps {
-  GUP: Record<string, string>;
+  params: Params;
   hideOverlay: () => void;
   showOverlay: () => void;
 }
 
 const App = (props: IAppProps): JSXInternal.Element => {
-  const skipChapterIntro = Boolean(props.GUP["skipChapterIntro"]);
+  const skipChapterIntro = props.params.getBoolean("skipChapterIntro", false);
 
   const [isGameReady, setGameReady] = useState<boolean>(false);
   const [isGameStartedState, setGameStartedState] = useState<boolean>(false);
@@ -115,7 +91,7 @@ const App = (props: IAppProps): JSXInternal.Element => {
 };
 
 export class FrontendUI {
-  constructor(sceneContainer: SceneContainer, globalLibs: IGlobalLibs, GUP: Record<string, string>) {
+  constructor(sceneContainer: SceneContainer, globalLibs: IGlobalLibs, params: Params) {
     const overlay = document.querySelector("#overlay");
     const showOverlay = () => {
       overlay.classList.remove("d-none");
@@ -128,7 +104,7 @@ export class FrontendUI {
     console.log("[FrontendUI] Rendering frontend App ...", globalLibs.preact.render);
     globalLibs.preact.render(
       <App
-        GUP={GUP}
+        params={params}
         sceneContainer={sceneContainer}
         globalLibs={globalLibs}
         hideOverlay={hideOverlay}