+++ src/ts/dom/frontend/TextCutsceneUI.tsx
@@ -10,6 +10,8 @@ import { useEffect, useState } from "preact/hooks";
 import { JSXInternal } from "preact/src/jsx";
 import { IFrontendUIBaseProps } from "./FrontendUI";
 import { useInterval } from "../customHooks";
+import { Typewriter } from "./Typewriter";
+import { MediaStorage } from "../../io/MediaStorage";
 
 interface TextCutsceneUIProps extends IFrontendUIBaseProps {
   text: string;
@@ -17,19 +19,47 @@ interface TextCutsceneUIProps extends IFrontendUIBaseProps {
 }
 
 export const TextCutsceneUI = (props: TextCutsceneUIProps): JSXInternal.Element => {
-  const [counter, setCounter] = useState<number>(0);
+  const [textSegmentCounter, setCounter] = useState<number>(0);
+  const [audio, setAudio] = useState<HTMLAudioElement | null>(null);
 
-  const clearInterval = useInterval(() => {
-    console.log("Counter");
-    setCounter(counter + 20);
-  }, 50);
+  // const audio = MediaStorage.getAudio("resources/audio/custom-sounds/typewriter-01.wav");
+  // audio.loop = true;
+  MediaStorage.getAudio("resources/audio/custom-sounds/typewriter-01.wav")
+    .then(fetchedAudio => {
+      console.log("audio fetched");
+      fetchedAudio.loop = true;
+      setAudio(fetchedAudio);
+    })
+    .catch(error => {
+      console.error("[TextCutsceneUI] Failed to load typewriter audio effect.", error);
+    });
 
-  useEffect(() => {
-    if (counter > props.text.length) {
-      clearInterval();
-      props.onTerminated();
+  // Split text into paragraphs
+  const textSegments = props.text.split(/\n(\n+)/g).filter((text: string) => text.trim().length > 0);
+  console.log("textSegments", textSegments.length);
+
+  const increaseTextSegmentCounter = () => {
+    const hasTerminated = textSegmentCounter + 1 === textSegments.length;
+    setCounter(textSegmentCounter + 1);
+    if (hasTerminated) {
+      globalThis.setTimeout(props.onTerminated, 1500);
     }
-  }, [counter, clearInterval]);
+  };
 
-  return <div>{props.text.substring(0, counter)}</div>;
+  // return <div>{props.text}</div>;
+  // console.log("textSegmentCounter", textSegmentCounter);
+  return audio ? (
+    <div className="jsx-typewriter-wrapper">
+      {textSegments.map((text: string, index: number) => {
+        return (
+          <Typewriter
+            audio={audio}
+            text={text}
+            onTerminated={increaseTextSegmentCounter}
+            isRunning={textSegmentCounter === index}
+          />
+        );
+      })}
+    </div>
+  ) : null;
 };